tweak: remove old roll scripts
