--- scripts/governance/add/addSeries/addJuneSeries/README.md
@@ -1,5 +1,0 @@
-==== Add June Series ====
-The deployment of June series will be done in two stages, days apart. Each one requires multisig approval.
-
-1. Deploy fyToken and pools using `deployJuneSeries.network.sh`
-2. Activate series and roll strategies onto them using `activateJuneSeries.network.sh`
