+++ scripts/governance/add/addSeries/addJuneSeries/activateJuneSeries.arbitrum.sh
