+++ scripts/governance/add/addSeries/addJuneSeries/activateJuneSeries.mainnet.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
 
 set -eux
-RUN="npx hardhat run --network mainnet"
 HERE=$(dirname $0)
+export CONF=$PWD/$HERE/addJuneSeries.mainnet.config
+RUN="npx hardhat run --network localhost"
 
-$RUN $HERE/addJuneSeries-3.ts
-$RUN $HERE/addJuneSeries-3.ts
-$RUN $HERE/addJuneSeries-3.ts
+# $RUN $HERE/../../../../../tools/joinLoan.ts
+# $RUN $HERE/../../../../../tools/loadTimelock.ts
+# $RUN $HERE/../../../../../tools/advanceTimeToMaturity.ts
+# $RUN $HERE/../../../../../tools/advanceTimeThreeDays.ts
+
+$RUN $HERE/activateJuneSeries.ts
+# $RUN $HERE/activateJuneSeries.ts
+# $RUN $HERE/activateJuneSeries.ts
+
+# $RUN $HERE/../.././../tools/poolRollBalances.ts
