--- scripts/governance/add/addSeries/addJuneSeries/activateJuneSeries.rinkeby.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-set -eux
-RUN="npx hardhat run --network rinkeby"
-HERE=$(dirname $0)
-
-$RUN $HERE/addJuneSeries-3.ts
-$RUN $HERE/addJuneSeries-3.ts
-$RUN $HERE/addJuneSeries-3.ts
