--- scripts/governance/add/addSeries/addJuneSeries/steps.md
@@ -1,6 +1,0 @@
-1. Deploy series
-2. Deploy pools
-3. Orchestrate series - assetsAndSeries/orchestrateSeriesProposal
-4. Add series - assetsAndSeries/addSeriesProposal
-5. Init pools - assetsAndSeries/initPoolsProposal
-6. Roll strategies - strategies/rollStrategies
