--- scripts/governance/add/addSeries/addJuneSeries23/deployJuneSeries.mainnet.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-set -eux
-HERE=$(dirname $0)
-export CONF=$PWD/$HERE/addJuneSeries.mainnet.config
-RUN="npx hardhat run --network localhost"
-
-$RUN $HERE/../../../deploy/deployFYTokens.ts
-$RUN $HERE/../../../deploy/deployEulerPools.ts
-$RUN $HERE/../../../deploy/deployNonTVPools.ts
