--- scripts/governance/add/addSeries/addSeptSeries/activateSeptSeries.mainnet.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-set -eux
-HERE=$(dirname $0)
-export CONF=$PWD/$HERE/addSeptSeries.mainnet.config
-RUN="npx hardhat run --network localhost"
-
-$RUN $HERE/addSeptSeries-3.ts
-$RUN $HERE/addSeptSeries-3.ts
-$RUN $HERE/addSeptSeries-3.ts
