--- scripts/governance/add/addSeries/addSeptSeries/addSeptSeries.rinkeby.sh
@@ -1,14 +1,0 @@
-#!/bin/bash
-
-set -eux
-RUN="npx hardhat run --network rinkeby"
-HERE=$(dirname $0)
-
-$RUN $HERE/addSeptSeries-1.ts
-$RUN $HERE/addSeptSeries-2.ts
-
-$RUN $HERE/loadTimelock.ts
-
-$RUN $HERE/addSeptSeries-3.ts
-$RUN $HERE/addSeptSeries-3.ts
-$RUN $HERE/addSeptSeries-3.ts
