--- scripts/governance/add/addSeries/addSeptSeries/deploySeptSeries.mainnet.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-set -eux
-HERE=$(dirname $0)
-export CONF=$PWD/$HERE/addSeptSeries.mainnet.config
-RUN="npx hardhat run --network localhost"
-
-$RUN $HERE/../../deploy/deployFYTokens.ts # deploy fyTokens
-$RUN $HERE/../../deploy/deployPools.ts # deploy pools
